hw1/requirements.txt
lshash3==0.0.8
matplotlib==2.1.1
pandas==0.21.1